[FEATURE] Configuration Update (#110)
* Broke down Configuration into smaller, manageable parts.

* Changelog && Patch Bump
