+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "komada",
-  "version": "0.13.0",
+  "version": "0.13.1",
   "author": "Evelyne Lachance",
   "description": "Komada: Croatian for 'pieces', is a modular bot system including reloading modules and easy to use custom commands.",
   "main": "app.js",
